[Favorite] 페이징, 정렬 기능 추가 (id, 평점, 리뷰 개수)

+++ backend/src/main/java/com/example/Flicktionary/domain/movie/dto/MovieResponse.java
@@ -16,7 +16,7 @@ public class MovieResponse {
     private final String title;
     private final String posterPath;
     @NonNull
-    private final float averageRating;
+    private final double averageRating;
     @NonNull
     private final int ratingCount;
